backend/requirements.txt
asgiref==3.11.1
boto3==1.42.49
botocore==1.42.49
Django==6.0.2
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
gunicorn==25.1.0
jmespath==1.1.0
packaging==26.0
psycopg==3.3.2
psycopg-binary==3.3.2
pypdf==6.7.0
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
